Move CorrentWeather and UpcomingWeather to src/screens directory
