+++ App.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { View, StyleSheet, SafeAreaView } from 'react-native'
-import CurrentWeather from './src/components/CurrentWeather'
-import UpcomingWeather from './src/components/UpcomingWeather'
+import CurrentWeather from './src/screens/CurrentWeather'
+import UpcomingWeather from './src/screens/UpcomingWeather'
 
 const App = () => {
   return (
